proto: split key_format module
