+++ crates/proto/src/dnssec/openssl.rs
@@ -583,6 +583,7 @@ impl PublicKey for Rsa<'_> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::dnssec::decode_key;
     use crate::dnssec::test_utils::{hash_test, public_key_test};
 
     #[test]
@@ -634,5 +635,43 @@ mod tests {
         test_case(&[0x00, 0x7f, 0x00, 0x80], &[0x7f, 0x00, 0x80]);
         test_case(&[0x80, 0x00, 0x80], &[0x00, 0x80, 0x00, 0x80]);
         test_case(&[0xff, 0x00, 0x80], &[0x00, 0xff, 0x00, 0x80]);
+    }
+
+    #[test]
+    fn test_rsa_encode_decode_der() {
+        let algorithm = Algorithm::RSASHA256;
+        let key = RsaSigningKey::generate(algorithm).unwrap();
+        let der = key.encode_der().unwrap();
+        decode_key(&der, None, algorithm, KeyFormat::Der).unwrap();
+    }
+
+    #[test]
+    fn test_rsa_encode_decode_pem() {
+        let algorithm = Algorithm::RSASHA256;
+        let key = RsaSigningKey::generate(algorithm).unwrap();
+        let pem = key.encode_pem(None).unwrap();
+        decode_key(&pem, None, algorithm, KeyFormat::Pem).unwrap();
+
+        let encrypted = key.encode_pem(Some("test password")).unwrap();
+        decode_key(&encrypted, Some("test password"), algorithm, KeyFormat::Pem).unwrap();
+    }
+
+    #[test]
+    fn test_ec_encode_decode_der() {
+        let algorithm = Algorithm::ECDSAP256SHA256;
+        let key = EcSigningKey::generate(algorithm).unwrap();
+        let der = key.encode_der().unwrap();
+        decode_key(&der, None, algorithm, KeyFormat::Der).unwrap();
+    }
+
+    #[test]
+    fn test_ec_encode_decode_pem() {
+        let algorithm = Algorithm::ECDSAP256SHA256;
+        let key = EcSigningKey::generate(algorithm).unwrap();
+        let pem = key.encode_pem(None).unwrap();
+        decode_key(&pem, None, algorithm, KeyFormat::Pem).unwrap();
+
+        let encrypted = key.encode_pem(Some("test password")).unwrap();
+        decode_key(&encrypted, Some("test password"), algorithm, KeyFormat::Pem).unwrap();
     }
 }
